+++ tests/integration/PaymentMethodTest.php
@@ -291,6 +291,35 @@ class PaymentMethodTest extends Setup
         $this->assertSame($customer->id, $result->paymentMethod->customerId);
     }
 
+    public function testCreate_fromPayPalRefreshToken()
+    {
+        $customer = Braintree\Customer::createNoValidate();
+        $http = new HttpClientApi(Braintree\Configuration::$global);
+
+        $result = Braintree\PaymentMethod::create([
+            'customerId' => $customer->id,
+            'paypalRefreshToken' => 'PAYPAL_REFRESH_TOKEN',
+        ]);
+
+        $this->assertSame($customer->id, $result->paymentMethod->customerId);
+        $this->assertSame("B_FAKE_ID", $result->paymentMethod->billingAgreementId);
+    }
+
+    public function testCreate_fromPayPalRefreshTokenWithoutUpgrade()
+    {
+        $customer = Braintree\Customer::createNoValidate();
+        $http = new HttpClientApi(Braintree\Configuration::$global);
+
+        $result = Braintree\PaymentMethod::create([
+            'customerId' => $customer->id,
+            'paypalRefreshToken' => 'PAYPAL_REFRESH_TOKEN',
+            'paypalVaultWithoutUpgrade' => true,
+        ]);
+
+        $this->assertSame($customer->id, $result->paymentMethod->customerId);
+        $this->assertNull($result->paymentMethod->billingAgreementId);
+    }
+
     public function testCreate_fromUsBankAccountNonce()
     {
         $customer = Braintree\Customer::createNoValidate();